+++ R/permutation.R
@@ -72,7 +72,7 @@ print.ser_permutation <- function(x, ...) {
   invisible(x)
 }
 
-## fake summary (we dont really provide a summary,
+## fake summary (we don't really provide a summary,
 ## but summary produces now a reasonable result --- same as print)
 summary.ser_permutation <- function(object, ...)
   object