+++ tests/testthat/test-permuation_vector.R
@@ -1,3 +1,4 @@
+library(testthat)
 library(seriation)
 library(dendextend) ## Needed because it redefined all.equal for dendrograms
 
@@ -30,20 +31,31 @@ expect_identical(length(ser_permutation(ser_permutation(sp), 1:10)), 2L)
 context("permute")
 
 ## vector
-expect_identical(permute(1:10, ser_permutation(1:10)), 1:10)
+v <- structure(1:10, names = LETTERS[1:10])
+expect_identical(permute(v, ser_permutation(1:10)), v[1:10])
 expect_identical(permute(LETTERS[1:10], ser_permutation(1:10)), LETTERS[1:10])
-expect_identical(permute(1:10, ser_permutation(10:1)), 10:1)
+expect_identical(permute(v, ser_permutation(10:1)), v[10:1])
 expect_identical(permute(LETTERS[1:10], ser_permutation(10:1)), LETTERS[10:1])
 
-expect_error(permute(1:10, ser_permutation(1:11)))
+expect_error(permute(v, ser_permutation(1:11)))
 
 ## matrix
-m <- matrix(runif(9), ncol=3)
+m <- matrix(runif(9), ncol=3, dimnames = list(1:3, LETTERS[1:3]))
 expect_identical(permute(m, ser_permutation(1:3, 3:1)), m[,3:1])
 expect_identical(permute(m, ser_permutation(3:1, 3:1)), m[3:1,3:1])
 
 expect_error(permute(m, ser_permutation(1:10, 1:9)))
 expect_error(permute(m, ser_permutation(1:9, 1:11)))
+
+expect_identical(permute(m, ser_permutation(3:1, 3:1), margin = 1), m[3:1, ])
+expect_identical(permute(m, ser_permutation(3:1, 3:1), margin = 2), m[ , 3:1])
+expect_identical(permute(m, ser_permutation(3:1), margin = 1), m[3:1, ])
+expect_identical(permute(m, ser_permutation(3:1), margin = 2), m[, 3:1])
+
+## data.frame
+df <- as.data.frame(m)
+expect_identical(permute(df, ser_permutation(1:3, 3:1)), df[,3:1])
+expect_identical(permute(df, ser_permutation(3:1, 3:1)), df[3:1,3:1])
 
 ## dist
 d <- dist(matrix(runif(25), ncol=5))